Fix clipping.

+++ game/components/component_types.h
@@ -59,8 +59,8 @@ template<>
 inline void save_primitive ( ostream &lbfile, const location_t &t )
 {
      save_primitive<uint8_t>( lbfile, t.region );
-     save_primitive<uint8_t>( lbfile, t.x );
-     save_primitive<uint8_t>( lbfile, t.y );
+     save_primitive<int16_t>( lbfile, t.x );
+     save_primitive<int16_t>( lbfile, t.y );
      save_primitive<uint8_t>( lbfile, t.z );
 }
 
@@ -114,8 +114,8 @@ inline void load_primitive ( istream &lbfile, component_type &t )
 template<>
 inline void load_primitive( istream &lbfile, location_t &t ) {
     load_primitive<uint8_t>( lbfile, t.region );
-    load_primitive<uint8_t>( lbfile, t.x );
-    load_primitive<uint8_t>( lbfile, t.y );
+    load_primitive<int16_t>( lbfile, t.x );
+    load_primitive<int16_t>( lbfile, t.y );
     load_primitive<uint8_t>( lbfile, t.z );
 }
